% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_assumptions.R
\name{check_all_assumptions}
\alias{check_all_assumptions}
\title{Check All Linear Regression Assumptions}
\usage{
check_all_assumptions(model)
}
\description{
Check the linearity, homoscedasticity, residual_normality, influential_points, and multicollinearity assumptions of linear regression.
}
\examples{
data(trees)
tree_model <- lm(Volume ~ Height + Girth, data=trees)
check_all_assumptions(tree_model)

}
