% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoscedasticity.R
\name{check_homoscedasticity}
\alias{check_homoscedasticity}
\title{Check Homoscedasticity}
\usage{
check_homoscedasticity(model, include_resid_fitted = T)
}
\description{
Check homoscedasticity assumption of linear regression.
}
\examples{
data(trees)
tree_model <- lm(Volume ~ Height + Girth, data=trees)
check_homoscedasticity(tree_model)

}
