% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influential_points.R
\name{check_influential_points}
\alias{check_influential_points}
\title{Check Unduly Influential Points Assumption}
\usage{
check_influential_points(model, cd_threshold = 4/model$df.residual)
}
\description{
Check the no unduly influential points assumption of linear regression.
}
\examples{
data(trees)
tree_model <- lm(Volume ~ Height + Girth, data=trees)
check_influential_points(tree_model)

}
