% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearity.R
\name{check_linearity}
\alias{check_linearity}
\title{Check Linearity}
\usage{
check_linearity(model, is_SLR = NULL)
}
\description{
Check linearity assumption of linear regression.
}
\examples{
data(trees)
tree_model <- lm(Volume ~ Height + Girth, data=trees)
check_linearity(tree_model)

}
