% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicollinearity.R
\name{check_multicollinearity}
\alias{check_multicollinearity}
\title{Check Minimal Multicollinearity Assumption}
\usage{
check_multicollinearity(model)
}
\description{
Check the minimal multicollinearity assumption of linear regression.
}
\examples{
data(trees)
tree_model <- lm(Volume ~ Height + Girth, data=trees)
check_multicollinearity(tree_model)

}
