% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_normality.R
\name{check_residual_normality}
\alias{check_residual_normality}
\title{Check Residual Normality}
\usage{
check_residual_normality(model)
}
\description{
Check residual normality assumption of linear regression.
}
\examples{
data(trees)
tree_model <- lm(Volume ~ Height + Girth, data=trees)
check_residual_normality(tree_model)

}
