% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_selection_best_subset.R
\name{var_selection_best_subset}
\alias{var_selection_best_subset}
\title{Best subsets variable selection}
\usage{
var_selection_best_subset(
  model,
  response = NULL,
  criteria = "AIC",
  return_model = F
)
}
\description{
Perform best subsets variable selection
}
\examples{
data(trees)
tree_model <- lm(Volume ~ Height + Girth, data=trees)
var_selection_best_subset(tree_model, criteria='AIC', return_model=F)

}
