% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_selection_regularization.R
\name{var_selection_elastic_net}
\alias{var_selection_elastic_net}
\title{Elastic Net regularization variable selection}
\usage{
var_selection_elastic_net(model, response = NULL, show_plot = T)
}
\description{
Perform Elastic Net regularization variable selection
}
\examples{
data(trees)
tree_model <- lm(Volume ~ Height + Girth, data=trees)
var_selection_elastic_net(tree_model, response = 'Volume', show_plot=T)

}
