% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_selection_regularization.R
\name{var_selection_lasso}
\alias{var_selection_lasso}
\title{LASSO regularization variable selection}
\usage{
var_selection_lasso(model, response = NULL, show_plot = T)
}
\description{
Perform LASSO regularization variable selection
}
\examples{
data(trees)
tree_model <- lm(Volume ~ Height + Girth, data=trees)
var_selection_lasso(tree_model, response = 'Volume', show_plot=T)

}
