% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_selection_regularization.R
\name{var_selection_regularization}
\alias{var_selection_regularization}
\title{Regularization variable selection}
\usage{
var_selection_regularization(model, type = "lasso", show_plot = T)
}
\description{
Perform regularization variable selection: LASSO or Elastic Net
}
\examples{
data(trees)
tree_model <- lm(Volume ~ Height + Girth, data=trees)
var_selection_regularization(tree_model, type='lasso', show_plot=T)

}
