% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_selection_step.R
\name{var_selection_sequential}
\alias{var_selection_sequential}
\title{Sequential replacement variable selection}
\usage{
var_selection_sequential(model, criteria = "AIC", print_summary = T)
}
\description{
Perform sequential replacement variable selection
}
\examples{
data(trees)
tree_model <- lm(Volume ~ Height + Girth, data=trees)
var_selection_sequential(tree_model, criteria='AIC', print_summary=T)

}
