% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_selection_step.R
\name{var_selection_step}
\alias{var_selection_step}
\title{Stepwise variable selection}
\usage{
var_selection_step(model, direction, criteria = "AIC", print_summary = T)
}
\description{
Perform stepwise variable selection: forward, backward, or sequential replacement
}
\examples{
data(trees)
tree_model <- lm(Volume ~ Height + Girth, data=trees)
var_selection_step(tree_model, direction='forward', criteria='AIC', print_summary=T)

}
